centroid: force fp64 throughout the kmeans + pivoted-Cholesky pipeline
Eliminates four fp32 leaks that were masking precision under "fp32
noise" labels in the diagnostics:

  centroid/pivoted_cholesky.py
    gather_wfn_at_candidates: complex64 buffer + .astype(complex64)
                              downcast → complex128 throughout.
                              Matches load_wfns.* precision.
    prune_candidates_by_pivoted_cholesky: kw_j float32 → float64.
    build_gram_q0_via_loadwfns: kw float32 → float64.

  centroid/kmeans_isdf.py
    weighted_kmeans_jax: metric_tensor, jnp.linspace grid coords,
                        rho_flat, history-array dtypes all float32 →
                        float64. Existing tests build inputs at fp32
                        and the function honors the caller's dtype, so
                        no test changes; only main()'s defaults change.
    main(): rho_jax / avec_jax cast to float64 (was float32).

The downstream load_wfns / compute_pair_density_spin_traced /
compute_gram_q0_from_left_right path was already complex128 native, so
the fp32 leaks above were the only thing keeping us from honest fp64.

Effect on Si 4×4×4 M=2000, N_c=1000, 2×2 mesh (full-BZ pipeline):

  Quantity                 mixed (run 20)   fp64 (run 21)
  first pivot              1.264e-6         1.264e-6
  pivot #500               3.550e-13        3.513e-13
  pivot #1000              1.469e-15        1.464e-15
  biggest leftover         1.453e-15        1.423e-15

Identical to within roundoff because the dominant path was already
fp64; the fix is mostly correctness/honesty rather than a numerical
shift here. The residual floor at ~1.5e-15 is the genuine fp64 noise
floor for this 64-k-point Gram (per-pivot ~64·ε relative roundoff,
times d0max ≈ 1.3e-6, gives a floor near 1e-15) — not fp32 chaff.

Pytest: 20 / 20 pass on the touched modules. (Test-side inputs still
explicitly fp32 to exercise dtype-genericity; main pipeline is fp64.)

+++ src/centroid/kmeans_isdf.py
@@ -777,19 +777,24 @@ def weighted_kmeans_jax(
           f"(mesh={'|'.join(str(a) + '=' + str(mesh.shape[a]) for a in (mesh_axis if isinstance(mesh_axis, tuple) else (mesh_axis,))) if mesh is not None else 'single-device'})...")
 
     # Precompute metric tensor: G = avec^T @ avec. Distances inherit avec's units.
-    metric_tensor = jnp.array(avec.T @ avec, dtype=jnp.float32)
+    # fp64 throughout: kmeans positions are < 1 in magnitude (fractional
+    # coords) so fp32 has only ~7 digits of precision near the cell origin —
+    # adequate for clustering a 24³ grid but limits convergence tolerance and
+    # propagates into downstream pivoted-Cholesky if any kmeans output feeds
+    # the wfn-loading layer (which is fp64 throughout). Standardize on fp64.
+    metric_tensor = jnp.array(avec.T @ avec, dtype=jnp.float64)
     tolerance_sq = tolerance ** 2  # Compare squared distances for efficiency.
 
-    # Build grid of fractional positions (float32 to match centroids dtype).
+    # Build grid of fractional positions (fp64 to match metric_tensor).
     grid_x, grid_y, grid_z = rho_jax.shape
     X, Y, Z = jnp.meshgrid(
-        jnp.linspace(0, 1, grid_x, endpoint=False, dtype=jnp.float32),
-        jnp.linspace(0, 1, grid_y, endpoint=False, dtype=jnp.float32),
-        jnp.linspace(0, 1, grid_z, endpoint=False, dtype=jnp.float32),
+        jnp.linspace(0, 1, grid_x, endpoint=False, dtype=jnp.float64),
+        jnp.linspace(0, 1, grid_y, endpoint=False, dtype=jnp.float64),
+        jnp.linspace(0, 1, grid_z, endpoint=False, dtype=jnp.float64),
         indexing="ij",
     )
     positions = jnp.stack((X, Y, Z), axis=-1).reshape(-1, 3)
-    rho_flat = rho_jax.reshape(-1).astype(jnp.float32)
+    rho_flat = rho_jax.reshape(-1).astype(jnp.float64)
     n_points = positions.shape[0]
 
     print(f"Grid size: {grid_x}x{grid_y}x{grid_z} = {n_points} points")
@@ -865,8 +870,8 @@ def weighted_kmeans_jax(
     # returned an empty placeholder here; flip ``_TRACK_HISTORY = True`` to
     # turn it back on when debugging.
     _TRACK_HISTORY = False
-    history = (np.zeros((N_c, max_steps), dtype=np.float32)
-               if _TRACK_HISTORY else np.zeros((N_c, 0), dtype=np.float32))
+    history = (np.zeros((N_c, max_steps), dtype=np.float64)
+               if _TRACK_HISTORY else np.zeros((N_c, 0), dtype=np.float64))
     steps_taken = max_steps
 
     if mesh is not None:
@@ -1145,8 +1150,8 @@ def main():
     print(f"Lattice lengths: {lattice_lengths} Å "
           f"(spacing: {lattice_lengths / np.array(charge_density.shape)} Å)")
 
-    rho_jax = jnp.asarray(charge_density, dtype=jnp.float32)
-    avec_jax = jnp.asarray(avec_ang, dtype=jnp.float32)
+    rho_jax = jnp.asarray(charge_density, dtype=jnp.float64)
+    avec_jax = jnp.asarray(avec_ang, dtype=jnp.float64)
 
     # Sharding: build a 2-D mesh ('x','y') whenever possible. This is the
     # same mesh shape used by the gw_jax ISDF pipeline, which lets the
